LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/WebIDL/ecmascript-binding/global-immutable-prototype.any.serviceworker-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')
